fix(B-464): split conversation category model and schema and delete ConversationCategoryCreate
