+++ backend/app/schemas/__init__.py
@@ -1,5 +1,6 @@
 from app.schemas.admin_dashboard_stats import AdminDashboardStatsRead
 from app.schemas.app_config import AppConfigCreate, AppConfigRead
+from app.schemas.conversation_category import ConversationCategoryRead
 from app.schemas.message_schema import MessageCreateSchema, MessageSchema
 
 __all__ = [
@@ -8,4 +9,5 @@ __all__ = [
     'AdminDashboardStatsRead',
     'AppConfigCreate',
     'AppConfigRead',
+    'ConversationCategoryRead',
 ]
